Unified options used for single crop from center
Until now, if you wanted to crop a region around the pixel (A,B) in an
image with Crop, you had to use `--xc=A --yc=B'. On the other hand, if you
wanted to a crop around the RA and Dec of (R,D), you have to use `--ra=R
--dec=D'. Each standard also has its own separate option to specify the
width: `--iwidth' and `--wwidth'. In parallel, Crop also has the `--mode'
option to specify the coordinate standard (for cases like `--polygon',
where only one option is used for both standards).

This can all be confusing, just adds to the number of options (makes things
harder for the user) and also makes maintaining Crop harder. The main
reason for this is historical: Crop is Gnuastro's first program, so I
wasn't too experience and all these options grew very gradually.

This commit addresses these issues as described below. Since these changes
were significant, they have also been included in the NEWS file so users
downloading the source before the 0.4 version can be aware of these major
changes in an old program.

  The new `--width' option is now used to define the width of a single
  crop. Hence the old `--iwidth', `--wwidth' were removed. The units to
  interpret the value to the option are specified by the `--mode'
  option. With the new `--width' option it is also possible to define a
  non-square crop (different widths along each dimension). In WCS mode, its
  units are no longer arcseconds but are the same units of the WCS (degrees
  for angles). `--width' can also accept fractions. So to set a width of 5
  arcseconds, you can give it a value of `5/3600' for the angular
  dimensions.

  The new `--center' option is now used to define the center of a single
  crop. Hence the old `--ra', `--dec', `--xc', `--yc' have been
  removed. This new option can take multiple values (one value for each
  dimension). Fractions are also acceptable.

  The new `--coordcol' option is now used to determine the catalog columns
  that define coordinates. Hence the old `--racol', `--deccol', `--xcol',
  and `--ycol' have been removed. This new option can be called multiple
  times and the order of its calling will be used for the column containing
  the center in the respective dimension (in FITS format).

The following library functions have also been corrected.

  `gal_box_overlap' now works on data of any dimensionality and thus also
  needs the number of dimensions (elements in each input array).

  `gal_box_border_from_center' now accepts an array of coordinates as one
  argument and the number of dimensions as another. This allows it to work
  on any dimensionality.

  `gal_wcs_pixel_scale' now replaces the old `gal_wcs_pixel_scale_deg',
  since it doesn't only apply to degrees. The pixel scale units are defined
  by the units of the WCS.

+++ lib/gnuastro/dimension.h
@@ -65,8 +65,8 @@ gal_dimension_num_neighbors(size_t ndim);
 /************************************************************************/
 /********************          Coordinates         **********************/
 /************************************************************************/
-#define GAL_DIMENSION_FLT_TO_INT(FLT) ( (FLT)-(int)(FLT) > 0.5f  \
-                                        ? (int)(FLT)+1 : (int)(FLT) )
+#define GAL_DIMENSION_FLT_TO_INT(FLT) ( (FLT)-(long)(FLT) > 0.5f  \
+                                        ? (long)(FLT)+1 : (long)(FLT) )
 
 void
 gal_dimension_add_coords(size_t *c1, size_t *c2, size_t *out, size_t ndim);
